pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Aleksandr Ivanov <Aleksandr200411y@.mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.8.1"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
tenacity = "^9.0.0"
alembic = "^1.15.1"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
python-jose = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
